add new Edge class as a topological base class for 1D elements and derive TemplateLine class from the class
