+++ MeshLib/Elements/TemplateLine.h
@@ -17,9 +17,11 @@
 
 #include <array>
 #include <limits>
+#include <cmath>
 
 #include "MeshEnums.h"
 #include "Element.h"
+#include "Edge.h"
 #include "Node.h"
 
 #include "MathTools.h"
@@ -33,51 +35,24 @@ namespace MeshLib {
  *  0--------1
  * @endcode
  */
-template<unsigned NNODES, CellType CELLEDGETYPE>
-class TemplateLine : public Element
+template<unsigned NNODES, CellType CELLLINETYPE>
+class TemplateLine : public Edge
 {
 public:
 	/// Constructor with an array of mesh nodes.
 	TemplateLine(Node* nodes[NNODES], unsigned value = 0);
 
-	/// Constructs an edge from array of Node pointers.
+	/// Constructs a line from array of Node pointers.
 	TemplateLine(std::array<Node*, NNODES> const& nodes, unsigned value = 0);
 
 	/// Copy constructor
-	TemplateLine(const TemplateLine &edge);
+	TemplateLine(const TemplateLine &line);
 
 	/// Destructor
 	virtual ~TemplateLine();
 
-	/// Returns the length, area or volume of a 1D, 2D or 3D element
-	double getContent() const { return _length; };
-
-	/// Returns the edge i of the element.
-	const Element* getEdge(unsigned i) const { (void)i; return NULL; };
-
-	/// Returns the face i of the element.
-	const Element* getFace(unsigned i) const { (void)i; return NULL; };
-
 	/// Compute the minimum and maximum squared edge length for this element
 	void computeSqrEdgeLengthRange(double &min, double &max) const { min = _length; max = _length; };
-
-	/// 1D elements have no edges
-	unsigned getNEdges() const { return 0; };
-
-	/// Get the number of nodes for face i.
-	unsigned getNFaceNodes(unsigned i) const { (void)i; return 0; };
-
-	/// Get the number of faces for this element.
-	unsigned getNFaces() const { return 0; };
-
-	/// Get the length of this 1d element.
-	double getLength() const { return _length; };
-
-	/// Get dimension of the mesh element.
-	unsigned getDimension() const { return 1; };
-
-	/// Get the number of neighbors for this element.
-	unsigned getNNeighbors() const { return 0; };
 
 	/// Get the number of nodes for this element.
 	virtual unsigned getNNodes(bool all = false) const
@@ -95,19 +70,15 @@ public:
 	 * Get the type of the element in context of the finite element method.
 	 * @return a value of the enum FEMElemType::type
 	 */
-	virtual CellType getCellType() const { return CELLEDGETYPE; }
+	virtual CellType getCellType() const { return CELLLINETYPE; }
 
-	/// Returns true if these two indices form an edge and false otherwise
-	bool isEdge(unsigned idx1, unsigned idx2) const
-	{
-		if (0==idx1 && 1==idx2) return true;
-		if (1==idx1 && 0==idx2) return true;
-		return false;
-	}
-
+	/**
+	 * Method clone is inherited from class Element. It makes a deep copy of the TemplateLine instance.
+	 * @return an exact copy of the object
+	 */
 	virtual Element* clone() const
 	{
-		return new TemplateLine<NNODES,CELLEDGETYPE>(*this);
+		return new TemplateLine<NNODES,CELLLINETYPE>(*this);
 	}
 
 	/**
@@ -128,17 +99,6 @@ protected:
 	{
 		return sqrt(MathLib::sqrDist(_nodes[0]->getCoords(), _nodes[1]->getCoords()));
 	}
-
-	/// 1D elements have no edges.
-	Node* getEdgeNode(unsigned edge_id, unsigned node_id) const { (void)edge_id; (void)node_id; return NULL; };
-
-	/// 1D elements have no faces.
-	Node* getFaceNode(unsigned face_id, unsigned node_id) const { (void)face_id; (void)node_id; return NULL; };
-
-	/// Returns the ID of a face given an array of nodes (but is not applicable for edges!).
-	unsigned identifyFace(Node* [3]/*nodes[3]*/) const { return std::numeric_limits<unsigned>::max(); };
-
-	double _length;
 
 }; /* class */
